[styles] Fix type declaration collission on dev machines
