+++ packages/material-ui-styles/src/StylesProvider/index.d.ts
@@ -1,0 +1,1 @@
+export { default } from './StylesProvider';
