+++ packages/material-ui-styles/src/ThemeProvider/index.d.ts
@@ -1,0 +1,1 @@
+export { default } from './ThemeProvider';
