+++ packages/material-ui-styles/src/createGenerateClassName/createGenerateClassName.d.ts
@@ -1,0 +1,11 @@
+import { GenerateClassName } from 'jss';
+
+export interface GenerateClassNameOptions {
+  dangerouslyUseGlobalCSS?: boolean;
+  productionPrefix?: string;
+  seed?: string;
+}
+
+export default function createGenerateClassName(
+  options?: GenerateClassNameOptions,
+): GenerateClassName;
