+++ packages/material-ui-styles/src/createGenerateClassName/index.d.ts
@@ -1,0 +1,1 @@
+export { default } from './createGenerateClassName';
