+++ packages/material-ui-styles/src/createStyled/index.d.ts
@@ -1,0 +1,1 @@
+export { default } from './createStyled';
