+++ packages/material-ui-styles/src/createStyles/index.d.ts
@@ -1,0 +1,1 @@
+export { default } from './createStyles';
