+++ packages/material-ui-styles/src/install/index.d.ts
@@ -1,0 +1,1 @@
+export { default } from './install';
