+++ packages/material-ui-styles/src/install/install.d.ts
@@ -1,0 +1,1 @@
+export default function install(): void;
