+++ packages/material-ui-styles/src/jssPreset/index.d.ts
@@ -1,0 +1,1 @@
+export { default } from './jssPreset';
