+++ packages/material-ui-styles/src/jssPreset/jssPreset.d.ts
@@ -1,0 +1,3 @@
+import { JSSOptions } from 'jss';
+
+export default function jssPreset(): JSSOptions;
