+++ packages/material-ui-styles/src/makeStyles/index.d.ts
@@ -1,0 +1,1 @@
+export { default } from './makeStyles';
