+++ packages/material-ui-styles/src/makeStyles/makeStyles.d.ts
@@ -1,0 +1,12 @@
+import {
+  ClassKeyOfStyles,
+  ClassNameMap,
+  PropsOfStyles,
+  Styles,
+  WithStylesOptions,
+} from '@material-ui/styles/withStyles';
+
+export default function makeStyles<S extends Styles<any, any>>(
+  styles: S,
+  options?: WithStylesOptions<ClassKeyOfStyles<S>>,
+): (props: PropsOfStyles<S>) => ClassNameMap<ClassKeyOfStyles<S>>;
