+++ packages/material-ui-styles/src/styled/index.d.ts
@@ -1,0 +1,1 @@
+export { default, StyledProps } from './styled';
