+++ packages/material-ui-styles/src/useTheme/index.d.ts
@@ -1,0 +1,1 @@
+export { default } from './useTheme';
