+++ packages/material-ui-styles/src/useTheme/useTheme.d.ts
@@ -1,0 +1,1 @@
+export default function useTheme<T>(): T;
