+++ packages/material-ui-styles/src/withStyles/index.d.ts
@@ -1,0 +1,2 @@
+export { default } from './withStyles';
+export * from './withStyles';
