+++ packages/material-ui-styles/src/withTheme/index.d.ts
@@ -1,0 +1,1 @@
+export { default, WithTheme } from './withTheme';
